+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## v1.2.0
+* When given a symlink, crun will attempt to follow/use it
+* Lower required Node engine to 8 from 10
+* Upgrade commander from 2.19.0 to 2.20.0
+* Change config option from using options to new signature, use `crun config --help` to view it.
+* Remove `-f` option for `start|stop|unload` and just implicitly check if using a path or not
+
 ## v1.1.0
 * If the cog.json does not specify a run command, check if there exists a package.json and then
     assume it is using node and the script under "main" in package.json.